fix: unit tests for account configs
