+++ apps/account-api/src/api.module.ts
@@ -5,7 +5,6 @@ import { ScheduleModule } from '@nestjs/schedule';
 import { BullBoardModule } from '@bull-board/nestjs';
 import { BullMQAdapter } from '@bull-board/api/bullMQAdapter';
 import { ExpressAdapter } from '@bull-board/express';
-import { ConfigModule } from '#account-lib/config/config.module';
 import { BlockchainModule } from '#account-lib/blockchain/blockchain.module';
 import { EnqueueService } from '#account-lib/services/enqueue-request.service';
 import { QueueModule } from '#account-lib/queues';
@@ -20,7 +19,7 @@ import {
 } from './controllers';
 import { AccountsService, HandlesService, DelegationService, KeysService } from './services';
 import { DelegationsControllerV2 } from './controllers/v2/delegation-v2.controller';
-import { ConfigModule as NestConfigModule } from '@nestjs/config';
+import { ConfigModule } from '@nestjs/config';
 import { allowReadOnly } from '#account-lib/blockchain/blockchain.config';
 import cacheConfig, { ICacheConfig } from '#account-lib/cache/cache.config';
 import queueConfig from '#account-lib/queues/queue.config';
@@ -28,8 +27,7 @@ import apiConfig from './api.config';
 
 @Module({
   imports: [
-    ConfigModule.forRoot(true), // allowReadOnly
-    NestConfigModule.forRoot({ isGlobal: true, load: [apiConfig, allowReadOnly, cacheConfig, queueConfig] }),
+    ConfigModule.forRoot({ isGlobal: true, load: [apiConfig, allowReadOnly, cacheConfig, queueConfig] }),
     BlockchainModule,
     EventEmitterModule.forRoot({
       // Use this instance throughout the application
@@ -50,7 +48,7 @@ import apiConfig from './api.config';
       ignoreErrors: false,
     }),
     CacheModule.forRootAsync({
-      imports: [NestConfigModule],
+      imports: [ConfigModule],
       useFactory: (cacheConf: ICacheConfig) => [
         {
           url: cacheConf.redisUrl.toString(),
