+++ libs/account-lib/src/cache/cache.config.ts
@@ -11,7 +11,10 @@ export default registerAs('cache', (): ICacheConfig => {
   const configs: JoiUtils.JoiConfig<ICacheConfig> = {
     redisUrl: {
       value: process.env.REDIS_URL,
-      joi: Joi.string().uri().required(),
+      joi: Joi.string()
+        .uri()
+        .required()
+        .custom((v) => new URL(v)),
     },
     cacheKeyPrefix: {
       value: process.env.CACHE_KEY_PREFIX,
