+++ package.json
@@ -88,6 +88,7 @@
     "lint:content-publishing": "eslint apps/content-publishing*/**/*.[tj]s libs/content-publishing-lib/**/*.[tj]s",
     "lint:content-watcher": "eslint apps/content-watcher/**/*.[tj]s libs/content-watcher-lib/**/*.[tj]s",
     "lint:graph": "eslint apps/graph*/**/*.[tj]s libs/graph-lib/**/*.[tj]s",
+    "lint:libs": "eslint libs/**/*.[tj]s",
     "lint:fix": "eslint --fix apps/**/*.[tj]s libs/**/*.[tj]s",
     "setup:content-publishing:chain": "node scripts/content-publishing/local-init.cjs",
     "setup:content-watcher:chain": "node scripts/content-watcher/local-chain-setup.cjs",
